+++ SKILL.md
@@ -1,6 +1,6 @@
 ---
 name: open-gis
-description: "Use this skill for production GIS/geospatial work, open-first but pragmatic about hosted/SaaS services when scale or data quality requires them: spatial data pipelines; vector/raster/point-cloud processing; satellite/EO imagery; LiDAR; CRS/projection/EPSG troubleshooting; spatial joins, buffers, distance/area analysis; spatial SQL / GeoSQL; routing, isochrones, geocoding; terrain/hydrology; tile generation; and web maps. Trigger when the user mentions GIS, geospatial, OpenStreetMap/OSM, Overture Maps, Sentinel, Landsat, STAC, LiDAR, GeoTIFF/COG, GeoParquet, Shapefile, GeoPackage, PMTiles, vector tiles, raster, CRS, EPSG, projections, WMS/WFS/WMTS/OGC API, QGIS, GDAL/OGR, GeoPandas, Shapely, xarray/rioxarray, DuckDB Spatial, PostGIS, BigQuery GIS, Snowflake geospatial, Sedona, PDAL, OSRM, Valhalla, GraphHopper, tippecanoe, Martin, MapLibre, Estonia data (Maa- ja Ruumiamet, ETAK, EPSG:3301/L-EST97), INSPIRE, or regional data portals. Do not trigger for simple location lookups, travel directions, or casual map references without analytical or production GIS work."
+description: "Use for production GIS and geospatial work: source discovery and provenance, vector/raster/point-cloud pipelines, CRS and metric analysis, spatial SQL, routing and isochrones, QGIS projects, tile generation, and web maps. Applies to tools and formats such as OSM, Overture, STAC, Sentinel/Landsat, LiDAR, GeoPackage, GeoParquet, COG, PMTiles, WMS/WFS/OGC APIs, GDAL, GeoPandas, DuckDB Spatial, PostGIS, QGIS, MapLibre, and Estonian spatial data including ETAK and EPSG:3301. Open-first, with hosted services when scale or reliability requires them. Do not use for casual map references, simple place lookups, or ordinary travel directions without analytical GIS work."
 ---
 
 # Open GIS Toolkit
@@ -53,6 +53,13 @@ Hard rules for every material analysis:
 * **Runs are rerunnable without the chat.** A fresh environment with the documented sources must reproduce the project. The conversation transcript is not part of the analytical dependency graph.
 * **Validation is a pipeline stage**, not prose advice, with machine-readable results.
 * **Separate analysis semantics from rendering.** Define semantic presentation roles; never reinvent layout/colors/UX arbitrarily on each run.
+* **Prove semantic predicates from data.** If the question depends on a property such as municipal ownership, active status, public access, accessibility, or legal designation, the selected source must expose an authoritative field or documented mapping for that property. Preserve unknown as unknown; never default a missing value to the desired class. Record and validate the exact predicate.
+* **Bounded APIs must prove completeness.** Record `numberMatched`/equivalent and page until all records are returned. A response whose count equals the request limit is incomplete until proven otherwise. Prefer server-side spatial and attribute filters, and validate matched-versus-returned counts.
+* **Overrides must be executable and verified.** Attribute/geometry overrides must target a real source feature and, for attribute changes, match the asserted prior value. Evidence must be non-placeholder and inspectable. Validation must report each override as applied, rejected, or not testable; merely listing an override is not application. Scenario geometry must be labeled hypothetical and kept visually/provenance-distinct from authoritative layers.
+* **One canonical implementation creates every declared output.** Convenience/E2E entrypoints may wrap `pipeline.py`, but must not duplicate its processing or QGIS generation logic. A clean-room run must prove that every manifest and QGIS datasource is created by that canonical path.
+* **QGIS success means valid layers, not a successful process exit.** Use project-relative sources, pin any container/runtime version, validate every local datasource and layer-tree ID, and when PyQGIS is available load the written project and require every layer `isValid()`. If runtime loading cannot be tested, record `not_testable`; never convert it to a pass.
+* **Manifest, report, and run metadata must agree.** Every declared required/domain check appears exactly once in the report; warnings/not-testable checks propagate to run/project status; run IDs and hashes match an actual `runs/*.json` record.
+* **Labels must match the operation.** Do not label Euclidean buffers as network walking/driving catchments. Use explicit terms such as “2 km straight-line proxy” unless an actual routable network/isochrone was computed.
 * Cheat-sheet: `references/project-spec.md` defines the full schema; `templates/` gives ready scaffolds; `examples/tartu-development` is a worked reference project matching the acceptance scenario.
 
 ## Modules — read the relevant reference(s) before starting work
