+++ examples/tartu-development/runs/run-20260825-123752.json
@@ -1,0 +1,59 @@
+{
+  "run_id": "run-20260825-123752",
+  "started_at": "2026-08-25T12:37:52.970839+00:00",
+  "completed_at": "2026-08-25T12:37:57.717550+00:00",
+  "status": "warning",
+  "inputs_hash": "sha256:64612775f56e8b1e066cf1747cb8362c5181c16fc3d5dbb6e32becae58fa002b",
+  "outputs_hash": "sha256:c2fc11d74406f2b480ff1cccc79eeed1af1fd3026c9f0735bb3a30f337ed32fd",
+  "source_manifest": "data/source/manifest.json",
+  "validation_report": "validation/latest-report.json",
+  "sources": [
+    {
+      "key": "cadastral_parcels",
+      "sha256": "sha256:c5e2024057c0b125f284d64f31efda247e58a87b00aed577f5acec1713df0373"
+    },
+    {
+      "key": "roads",
+      "sha256": "sha256:58170565ec51eaa9f50f2a643e2224bcc1ce039e394ccb499f426e529859323f"
+    },
+    {
+      "key": "education_pois",
+      "sha256": "sha256:0cc7e785a7444a17961e6ff84878fa4bc3775d140f065ea1ea2d0d98d43e2ddd"
+    }
+  ],
+  "outputs": [
+    {
+      "path": "data/derived/final-candidates.gpkg",
+      "sha256": "sha256:a7c261c85010c8a1ec5f8b695fbb89560bfb3c89b402c6e765dcaff91fdb839e"
+    },
+    {
+      "path": "data/derived/final-candidates.parquet",
+      "sha256": "sha256:774108ab18fbe1e4e928e3b7e54c485dbe5ee441626ba48e86f531a2b87af9f0"
+    },
+    {
+      "path": "data/derived/final-candidates.json",
+      "sha256": "sha256:484d0ae9658ed8af1ad24b1149584ff69297d44e8d150190952a6651c0aa3b12"
+    },
+    {
+      "path": "data/derived/education_catchments.json",
+      "sha256": "sha256:a6cb3ea96baa987fa74378f3b5cfb2d51abeb964864c1514aa9cb8167dfae07b"
+    },
+    {
+      "path": "data/derived/education_pois.json",
+      "sha256": "sha256:0cc7e785a7444a17961e6ff84878fa4bc3775d140f065ea1ea2d0d98d43e2ddd"
+    },
+    {
+      "path": "data/derived/main_roads.json",
+      "sha256": "sha256:db8d744ae50c31ecebb8e307db6b0b19cb09598f532ff3356ff5d096e057d645"
+    },
+    {
+      "path": "project.qgz",
+      "sha256": "sha256:d1c3c680076fd409cf0c90b14cd21ae136123e35442e60336bb2e6f82929b058"
+    }
+  ],
+  "environment": {
+    "python": "3.12.3",
+    "duckdb": "1.5.5",
+    "pyproj": "3.7.2"
+  }
+}
